feat: refactors for heroku deploy, adds difference between envs
