+++ packages/fm-os-gql/src/Procfile
@@ -1,0 +1,1 @@
+web: node ./dist/index.js
