--- packages/fm-os-gql/src/config.sample.js
@@ -1,5 +1,0 @@
-// @see https://developer.github.com/v4/guides/forming-calls/#authenticating-with-graphql
-export const GITHUB_AUTH_TOKEN =
-  'CREATE YOUR OWN API KEY TO GET ACCESS TO GITHUB GRAPHQL API'
-
-export const FIREBASE_SERVICE_ACCOUNT = {}
