+++ packages/fm-os-gql/src/scripts/env.js
@@ -1,0 +1,4 @@
+if (process.env.NODE_ENV !== 'production') {
+  require('dotenv').config()
+  console.info('🎚️ Development env variables have been set!')
+}
